clean up type definitions
