+++ src/types.ts
@@ -247,3 +247,14 @@ export interface Flag {
   color: string;
   officerDescription?: string;
 }
+
+export type ResultData = (
+  flagSet?: string,
+  overrides?: { [flagId: string]: { heading?: string; description?: string } }
+) => {
+  [category: string]: {
+    flag: Flag;
+    responses: any[];
+    displayText: { heading: string; description: string };
+  };
+}
