--- src/types/flags.ts
@@ -1,8 +1,0 @@
-export interface Flag {
-  category: string;
-  value?: string;
-  text: string;
-  bgColor: string;
-  color: string;
-  officerDescription?: string;
-}
